First Bot
This bot will search for #100DaysOfCode, #Coding, #Python tweets on twitter in a sequence. If a tweet is not retweeted it will do that after each five-second. If the tweet is already retweeted then it will wait for 3 sec. It likes a tweet only after four retweets are done on a single tag.